renaming the daas model and removing un-necessary imports
